profiles/way_handler: Add support for route_weight_factor:motor_vehicle
Currently the profile only depends on road class, maxspeed
and very few physical items of the road like lanes, lane_markings
etc.

There are corner cases where the preferred route leaves the
higher class network to use a shortcut on a rough road.

Depending on the relative cost/weight of the 2 routes its impossible
to fix just by physical tags.

Introduce an OpenStreetMap tag which is a factor for
the roads weight calculated. This allows to generically make
roads slower or faster as their physical tags imply.

route_weight_factor:motor_vehicle=-0.5

Would apply -50% weight reduction on the calculated roads weight
by adding to the penalty.

The value is clamped to -+0.8 to not make roads completely
inaccessible or boost to much.

profile has been teste in RouteQA and the tag has been introduced
to a handful of shortcuts which appeared as issues in RouteQA
in Northrhine-Westfalia/Germany.