Fix and reverse access tag processing assumptions for car profile
This patch adds parsing of ";" seperated access tags values e.g.
"forestry;agricultural" and reverses the assumption of
accessibility of unknown values.

Before OSRM did not split access tag values and assumed unknowns to
be okay to use. As the string "forestry;agricultural" was not in the
negative list it was okay to use.

Now we split access tags into seperate strings. If one is
in our positive list (yes, permissive) the way is good to use.

If we have any element in the restricted list its restricted.

If we have only unknowns the way it is inaccessible.

So ways with access values of "unknown", "foobar", "i;dont;know" will
start to be inaccessible.

Ways with "delivery;private", "delivery;customers" will be restricted,
"forestry;agricultural" will be inaccessible.